Return received value when it isn't checked via regex
This change was spurred by the fact that validators should not set
default values for columns. If a ResultSource was to specify a default
value for a column that uses _val_text for example, it would not assign
the default value because a blank string was returned instead. IE,
columns that are non-mandatory should be able to return undef.

This may cause breakage in some uses of this module because end users
may have relied on this default value for their databases. Instead
those users should set the default_value for those columns.